remove cli profiles
